Changed file structure for testing `setup.py`
