+++ source/app_copy_draggable.py
@@ -13,22 +13,21 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-# File: dndapp0.py
+# File: app_copy_draggable.py
 #       Simple example of the DragNDropWidget Kivy library.
-#       Exercises some basic features of the library: setting droppable_zone_objects,
-#       drop_ok_animation_time, drop_func and failed_drop_func.
+#       Expands on dndapp0.py. Additional features of the library:
+#       * Copy the draggable widget, instead of removing it from the parent.
+#       * Can drop onto the original's parent
+#       * The parent has a drop_func defined
 
 from __future__ import print_function
 
 from kivy.app import App
-from kivy.lang import Builder
-import draggablestuff
 from kivy.clock import Clock
+from kivy.lang import Builder
 from kivy.uix.label import Label
-from debug_print import debug_print, set_debug_flag
-set_debug_flag(False)
 
-# Classes in the FloatLayout are defined in DraggableButton.py, included above.
+# Classes in the FloatLayout are defined in draggablestuff.py, included above.
 kv = '''
 FloatLayout:
     DragSourceBoxLayout:
@@ -41,26 +40,17 @@ FloatLayout:
                 size: self.size
         size_hint: 0.8, 0.25
         pos_hint: {'x': 0.1, 'y': 0.4}
+
         DraggableButton:
             text: 'Button 1'
-            droppable_zone_objects: [upper_to_box]
+            droppable_zone_objects: [upper_to_box, from_box]
             drop_ok_animation_time: 1.5
             drop_func: app.greet
-            drop_args: [ root ]
+            drop_args: [ root, self ]
             failed_drop_func: app.oops
-            failed_drop_args: [ from_box.__self__, root ]
-            not_drop_ok_do_animation: False
-        # This is not draggable:
-        Label:
-            text: '[color=ff3333]Regular[/color][color=3333ff]Label[/color]'
-            markup: True
-            size_hint: None, None
-            canvas.before:
-                Color:
-                    rgb: 0.7, 0.7, 0.7, 0.6
-                Rectangle:
-                    pos: self.pos
-                    size: self.size
+            failed_drop_args: [ root, self ]
+            remove_on_drag: False
+            can_drop_into_parent: True
 
     DragDestinationLabel:
         id: upper_to_box
@@ -89,7 +79,6 @@ DialogLabel:
     pos_hint: {'x': 0.3, 'y': 0.65}
 '''
 
-
 class DialogLabel(Label):
 
     def __init__(self, *args, **kwargs):
@@ -101,8 +90,8 @@ class DialogLabel(Label):
         self.rgba_list_orig = self.rgba_list
         Clock.schedule_interval(self.cycle_color, 0.3)
 
-    def cycle_color(self, delta_time):
-        debug_print ("DialogLabel:cycle_color, parent:", self.parent)
+    def cycle_color(self, dt):
+        print (self.rgba_list)
         if self.i < 6:
             if self.toggle_color:
                 # toggled color
@@ -114,11 +103,9 @@ class DialogLabel(Label):
                 self.toggle_color = True
             self.i += 1
         else:
-            debug_print ("Me: ", self, "end of cycle_color")
             Clock.unschedule(self.cycle_color)
             self.i = 0
             self.toggle_color = True
-            debug_print ("Me: ", self, " My Parent:", self.get_parent_window())
             self.parent.remove_widget(self)
 
 class dndapp0(App):
@@ -128,36 +115,27 @@ class dndapp0(App):
     def build(self):
         return Builder.load_string(kv)
 
-    def greet(self, calling_widget, root_widget):
+    def greet(self, arg1=None, arg2=None):
         '''
         
-        :param calling_widget: This is the first parameter given to drop_func() from
-        on_successful_drop(). It is the widget that's being dropped.
-        :param root_widget: the root widget found from the Kivy language, above
+        :param arg1: the root window
+        :param arg2: the widget that calls this method
         :return: 
         '''
-        kv_root = root_widget
+        kv_root = arg1
         messagebox = Builder.load_string(kv1)
         messagebox.text = "Dragging done!!!"
 
-        debug_print ("App Greet: add messagebox")
-        root_widget.add_widget(messagebox)
+        kv_root.parent.add_widget(messagebox)
         messagebox.flash()
 
-    def oops(self, the_widget=None, parent=None, kv_root=None):
-        """
-
-        :param the_widget: The DragNDropWidget that was dragged and dropped.
-        :param parent: The parent of the DragNDropWidget that was dragged and dropped.
-        :param kv_root: The FloatLayout at the root of the kv language.
-        :return:
-        """
-
+    def oops(self, arg1=None, arg2=None):
+        print ("Self, arg1, arg2:", self, arg1, arg2)
         messagebox = Builder.load_string(kv1)
         messagebox.text = "Ooops! Can't drop there!"
 
-        debug_print ("App Oops: add messagebox", self)
-        kv_root.parent.add_widget(messagebox) # Add it to the Top-level Window
+        kv_root = arg1
+        kv_root.parent.add_widget(messagebox)
         messagebox.flash()
 
 if __name__ == '__main__':
