+++ source/dragndropwidget.py
@@ -19,14 +19,16 @@
 #       the drag and drop widget library for Kivy.
 from __future__ import print_function
 
-from kivy.core.window import Window
-from kivy.animation import Animation
 import copy
-from kivy.uix.widget import Widget
+
+from kivy.animation import Animation
+from kivy.core.window import Window
 from kivy.properties import (
-    ListProperty, NumericProperty, BooleanProperty, ObjectProperty, StringProperty)
-import traceback
-from debug_print import Debug
+	ListProperty, NumericProperty, BooleanProperty, ObjectProperty, StringProperty)
+from kivy.uix.widget import Widget
+
+from source.debug_print import Debug
+
 debug = Debug() # Is False by default.
 
 # This looks like:
