+++ source/relative_stuff.py
@@ -24,12 +24,13 @@ Created on Oct 24, 2012
 
 from __future__ import print_function
 
-from dragndropwidget import DragNDropWidget
+from kivy.clock import Clock
+from kivy.uix.boxlayout import BoxLayout
 from kivy.uix.button import Button
 from kivy.uix.label import Label
-from kivy.uix.boxlayout import BoxLayout
 from kivy.uix.relativelayout import RelativeLayout
-from kivy.clock import Clock
+
+from source.dragndropwidget import DragNDropWidget
 
 
 class DraggableButton(Button, DragNDropWidget):
